+++ CHANGELOG.md
@@ -10,6 +10,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 ### Added
 
+- Add rule "double-words"
 - Add option `--punc-ignore-ellipsis`
 - Add rules "punc-space-id" and "punc-space-str"
 - Add rule "urls"